Refactoring settings screen: activity to fragment
