+++ app/src/main/res/layout/fragment_settings.xml
@@ -1,23 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    style="@style/SecondaryActivityStyle">
+    xmlns:tools="http://schemas.android.com/tools"
+    style="@style/SecondaryActivityStyle"
+    tools:context="ui.settings.SettingsFragment">
 
-    <LinearLayout
-        android:layout_width="match_parent"
+    <TextView
+        style="@style/SecondaryActivityTitleStyle"
         android:layout_height="wrap_content"
-        android:padding="4dp">
-
-        <ImageView
-            android:id="@+id/backButton"
-            style="@style/BackButtonNavigationBarStyle" />
-
-        <TextView
-            style="@style/SecondaryActivityTitleStyle"
-            android:layout_height="match_parent"
-            android:layout_marginStart="12dp"
-            android:text="@string/settings_button" />
-
-    </LinearLayout>
+        android:layout_marginVertical="12dp"
+        android:layout_marginStart="12dp"
+        android:text="@string/settings_button" />
 
     <FrameLayout
         style="@style/SettingsItemStyle"
